+++ oraclerdbms/resource_user_test.go
@@ -17,6 +17,15 @@ func TestAccUser(t *testing.T) {
 					resource.TestCheckResourceAttr("oraclerdbms_user.test", "username", "USER666"),
 				),
 			},
+			resource.TestStep{
+				Config: testAccUserConfigBasicUpdate,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("oraclerdbms_user.test", "default_tablespace", "SYSTEM"),
+					resource.TestCheckResourceAttr("oraclerdbms_user.test", "account_status", "LOCKED"),
+					resource.TestCheckResourceAttr("oraclerdbms_user.test", "quota.%", "1"),
+					resource.TestCheckResourceAttr("oraclerdbms_user.test", "quota.SYSTEM", "10M"),
+				),
+			},
 		},
 	})
 }
@@ -33,19 +42,33 @@ func TestAccUser_importBasic(t *testing.T) {
 			},
 
 			{
-				ResourceName:            resourceName,
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"password"},
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})
 }
 
-const testAccUserConfigBasic = `
+const (
+	testAccUserConfigBasic = `
 resource "oraclerdbms_user" "test" {
 	username = "USER666"
-	password = "change_on_install"
 	default_tablespace = "USERS"
+	account_status = "OPEN"
+	quota {
+		SYSTEM = "10M"
+	}
 }
 `
+	testAccUserConfigBasicUpdate = `
+resource "oraclerdbms_user" "test" {
+	username = "USER666"
+	default_tablespace = "SYSTEM"
+	account_status = "LOCKED"
+	quota {
+		SYSTEM = "10M"
+	}
+}
+`
+)